+++ core/src/commonMain/kotlin/internal/packets/ChannelPackets.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.Optional
 import kotlinx.serialization.Serializable
 
 internal interface ChannelPacket : EntityPacket {
-    val type: Int
+    val type: Byte
 }
 
 internal interface TextChannelPacket : ChannelPacket {
@@ -20,7 +20,7 @@ internal interface TextChannelPacket : ChannelPacket {
 
 internal interface GuildChannelPacket : ChannelPacket {
     var guild_id: Long?
-    val position: Int
+    val position: Short
     val name: String
     val nsfw: Boolean
     val permission_overwrites: List<PermissionOverwritePacket>
@@ -30,9 +30,9 @@ internal interface GuildChannelPacket : ChannelPacket {
 @Serializable
 internal data class GuildTextChannelPacket(
     override val id: Long,
-    override val type: Int,
+    override val type: Byte,
     @Optional override var guild_id: Long? = null,
-    override val position: Int,
+    override val position: Short,
     override val permission_overwrites: List<PermissionOverwritePacket>,
     override val name: String,
     @Optional val topic: String? = null,
@@ -46,33 +46,33 @@ internal data class GuildTextChannelPacket(
 @Serializable
 internal data class GuildVoiceChannelPacket(
     override val id: Long,
-    override val type: Int,
+    override val type: Byte,
     override var guild_id: Long?,
-    override val position: Int,
+    override val position: Short,
     override val permission_overwrites: List<PermissionOverwritePacket>,
     override val name: String,
     @Optional override val nsfw: Boolean = false,
     val bitrate: Int,
-    val user_limit: Int,
+    val user_limit: Byte,
     @Optional override val parent_id: Long? = null
 ) : GuildChannelPacket
 
 @Serializable
 internal data class GuildChannelCategoryPacket(
     override val id: Long,
-    override val type: Int,
+    override val type: Byte,
     @Optional override var guild_id: Long? = null,
     override val name: String,
     @Optional override val parent_id: Long? = null,
     @Optional override val nsfw: Boolean = false,
-    override val position: Int,
+    override val position: Short,
     override val permission_overwrites: List<PermissionOverwritePacket>
 ) : GuildChannelPacket
 
 @Serializable
 internal data class DmChannelPacket(
     override val id: Long,
-    override val type: Int,
+    override val type: Byte,
     val recipients: List<UserPacket>,
     @Optional override val last_message_id: Long? = null,
     @Optional override val last_pin_timestamp: String? = null
@@ -81,7 +81,7 @@ internal data class DmChannelPacket(
 @Serializable
 internal data class GroupDmChannelPacket(
     override val id: Long,
-    override val type: Int,
+    override val type: Byte,
     val owner_id: Long,
     val name: String,
     val icon: String,
@@ -93,31 +93,31 @@ internal data class GroupDmChannelPacket(
 @Serializable
 internal data class GenericChannelPacket(
     val id: Long,
-    val type: Int,
+    val type: Byte,
     @Optional val guild_id: Long? = null,
-    @Optional val position: Int? = null,
+    @Optional val position: Short? = null,
     @Optional val permission_overwrites: List<PermissionOverwritePacket>? = null,
     @Optional val name: String? = null,
     @Optional val topic: String? = null,
     @Optional val nsfw: Boolean = false,
     @Optional val last_message_id: Long? = null,
     @Optional val bitrate: Int? = null,
-    @Optional val user_limit: Int? = null,
+    @Optional val user_limit: Byte? = null,
     @Optional val recipients: List<UserPacket> = emptyList(),
     @Optional val icon: String? = null,
     @Optional val owner_id: Long? = null,
     @Optional val application_id: Long? = null,
     @Optional val parent_id: Long? = null,
     @Optional val last_pin_timestamp: String? = null,
-    @Optional val rate_limit_per_user: Int? = null
+    @Optional val rate_limit_per_user: Byte? = null
 )
 
 @Serializable
 internal class GenericTextChannelPacket(
     val id: Long,
-    val type: Int,
+    val type: Byte,
     @Optional val guild_id: Long? = null,
-    @Optional val position: Int? = null,
+    @Optional val position: Short? = null,
     @Optional val permission_overwrites: List<PermissionOverwritePacket>? = null,
     @Optional val name: String? = null,
     @Optional val topic: String? = null,
@@ -129,22 +129,22 @@ internal class GenericTextChannelPacket(
     @Optional val application_id: Long? = null,
     @Optional val parent_id: Long? = null,
     @Optional val last_pin_timestamp: String? = null,
-    @Optional val rate_limit_per_user: Int? = null
+    @Optional val rate_limit_per_user: Byte? = null
 )
 
 @Serializable
 internal data class GenericGuildChannelPacket(
     val id: Long,
-    val type: Int,
+    val type: Byte,
     @Optional val guild_id: Long? = null,
-    val position: Int,
+    val position: Short,
     @Optional val permission_overwrites: List<PermissionOverwritePacket> = emptyList(),
     @Optional val name: String? = null,
     @Optional val topic: String? = null,
     @Optional val nsfw: Boolean = false,
     @Optional val last_message_id: Long? = null,
     @Optional val bitrate: Int? = null,
-    @Optional val user_limit: Int? = null,
+    @Optional val user_limit: Byte? = null,
     @Optional val parent_id: Long? = null,
     @Optional val last_pin_timestamp: String? = null
 )
